requirements.txt
alabaster==0.7.12
attrs==21.4.0
Babel==2.10.1
certifi==2021.10.8
charset-normalizer==2.0.12
docutils==0.17.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
pyhere==1.0.0
pyparsing==3.0.8
pyprojroot==0.2.0
pytest==7.1.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
